requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
PyQt5>=5.15.0
statsmodels>=0.13.0
python-dotenv>=0.19.0
requests>=2.26.0
pyyaml>=6.0
pyinstaller>=5.0.0
scikit-learn>=1.0.0
pennylane>=0.22.0  # For quantum-inspired algorithms
shap>=0.40.0       # For explainable AI
geopandas>=0.10.0  # For geospatial analysis
folium>=0.12.0     # For interactive maps
plotly>=5.5.0      # For interactive visualizations
networkx>=2.6.0    # For network analysis of financial exposures
joblib>=1.1.0      # For parallel processing
tqdm>=4.62.0       # For progress bars
tensorflow>=2.8.0